Tracked mode cancels current movement if new one comes in
Allows direction reversal while a longer move was being executed.

+++ software/delta-control/Src/drivers/motion_types.c
@@ -114,7 +114,7 @@ int32_t cartesian_distance_between(CartesianPoint_t *a, CartesianPoint_t *b)
         int32_t delta_y = a->y - b->y;
         int32_t delta_z = a->z - b->z;
         float dist =  sqrtf( (delta_x*delta_x) + (delta_y*delta_y) + (delta_z*delta_z) ) ;
-        distance = abs( (uint32_t)dist );
+        distance = abs( dist );
     }
 
     return distance;
